libc: Merge test cases of VFS into CWD
There's currently no explicit VFS module in libc. This patch merges the
test case for VFS into the CWD test cases. The only VFS test case if for
`chdir_tx()`, which actually operates on CWD state.
